refactor: fetch 및 state 커스텀 훅으로 분리
